Fixed computation of metrics - now we get a bit fat dataframe with everything
the dataframe also includes the total number of core hours available.
Also, lib.v2.sh has been changed to use a different list of settings for the analysis
and for the runs, since analysis can be done on a larger set of runs than,
the run setup can.
For example, we can run the analysis on the results contained in
slurm_sim_data5/* and slurm_sim_data5_2/* by symlinking the results in
slurm_sim_data5_2 in slurm_sim_data5, but the 'run' settings must not include
the symlinked dirs from slurm_sim_data5_2, while the analysis settings should.